Completed ircservices conversion.
git-svn-id: http://svn.neuromancy.net/magick/trunk/Magick-II@1540 1d68a2a4-f605-0410-a2a0-863aa388aa54

+++ src/convert/epona.cpp
@@ -2106,7 +2106,7 @@ Nick_Stored_t *Convert::epona_CreateNickEntry(epona_NickAlias * na, epona_NickCo
     ETCB();
 }
 
-static mstring epona_getmodes(int16 modes)
+static mstring epona_getmodes(int32 modes)
 {
     BTCB();
     mstring retval;
